Added SifRebootIop (high level wrapper for SifResetIop).
Removed "SifSetReg(SIF_REG_SMFLAG, 0x40000)" from existing SifSyncIop and inserted it into SifResetIop.

This was done to ensure that multiple calls to SifSyncIop don't end up unable to return successfully when it should, it also fixes SifSyncIop returning a premature "success" result when the iop is in fact not ready yet.