Revert "Fix for the workaround for the Chinese SATA network adaptor: I forgot that"
This reverts commit 1b0f9a172b184ebf4f460ac119c8d19f947ce4aa.

I modified PS2Ident to show the entire 16-bit revision register, and it
reports its revision as 0x00FF. Rolling back to the previous revision
because that revision was working correctly.

It IS debatable whether it's worth keeping a workaround for this broken
adaptor, but I've decided that there's no harm in providing a workaround
for it. If someday, somebody releases a good SATA network adaptor, please
feel free to entirely remove this workaround!