Added a copy of the SCE workaround for iWakeupThread() on the EE.
Despite its documentation, iWakeupThread will error out if the specified thread is in THS_RUN state.
This workaround will determine if the running thread is the specified thread.
If it is not, then the old iWakeupThread syscall is invoked.
If it is, then the thread is first suspended, before iWakeupThread() is invoked. By suspending the thread, the thread enters THS_SUSPEND state, which is a state that iWakeupThread() will increment the wakeup request counter for.