Changed priorities of NETMAN threads: IOP Rx (IOP -> EE): 0x29 IOP Tx (EE -> IOP): 0x27 EE Rx (IOP -> EE): 0x59 EE Tx (EE -> IOP): 0x57
The idea is that the thread that passes frames to another thread, should have a higher priority than the thread it is passing frames too.
The only exception is the EE Tx thread. While changing the priorities for the other threads resulted in better performance, lowering the priority of the EE Tx thread worsened sending performance.
I guess that the TCP/IP stack causes the Tx thread to be delayed, so the Tx thread should have a higher priority.