(NETMAN) Replaced system for reducing the number of interrupt assertions that was introduced in da8b187: - SifSetReg/SifGetReg on the IOP are critical sections. - There was no way to prevent the EE-side from going to sleep after the IOP checks on its status, hence the implementation did not work well. - Reduced frame group size from 64 / 4 = 16 to just 8 because 16*2 = 32. Approximately 10 frames can be received in 1ms and having even only 31 free slots will prevent the call to sceSifSetDma from working.
With the new design, I hope that performance will be better because wisi indicated that the call to sceSifSetDma() on the IOP results in a substential delay because of its interrupt handler.
If the DMA channels (like SIF0) can manage 100MB/s, then it only takes about 15us to transfer a frame. There is additional overhead as well.
In 15us, the IOP can only manage 550 instructions. Since it must also wait for further frames over DMA channel 8 (DEV9), the SIF interrupt handler would also run once for every frame received.

Note: for now, the EE side does not share the new mechanism.
I haven't figured out how to make it work for the EE side (stack -> IOP), but I think it might not be necessary anyway - since the EE is so much faster and it does not have to wait for further frames via DMA (they're either already there or not).